cap nhat

+++ src/main/java/org/example/casestudymodule4shoestore/repositories/cart/ICartRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface ICartRepository extends JpaRepository <Cart,Integer> {
-    Cart findByCustomerId (Integer customerId);
+    Cart findByCustomerId (Long customerId);
 }
